+++ Modules/Registration/FPFH/include/itkPointFeature.hxx
@@ -81,7 +81,7 @@ template <typename TInputPointSet, typename TOutputPointSet>
 typename PointFeature<TInputPointSet, TOutputPointSet>::FeatureTypePointer
 PointFeature<TInputPointSet, TOutputPointSet>::ComputeSPFHFeature(TInputPointSet * input,
                                                                   TInputPointSet * input_normals,
-                                                                  unsigned int     radius,
+                                                                  double           radius,
                                                                   unsigned int     neighbors)
 {
   PointsLocatorTypePointer kdtree = PointsLocatorType::New();
@@ -195,7 +195,7 @@ template <typename TInputPointSet, typename TOutputPointSet>
 void
 PointFeature<TInputPointSet, TOutputPointSet>::ComputeFPFHFeature(TInputPointSet * input,
                                                                   TInputPointSet * input_normals,
-                                                                  unsigned int     radius,
+                                                                  double           radius,
                                                                   unsigned int     neighbors)
 {
   unsigned long int   num_of_points = input->GetNumberOfPoints();
